chore: converted ReferenceTableService to Kotlin
